requirements.txt
flasgger==0.9.7b2
Flask
paho_mqtt